feat: implemented autoPair option, true by default
When autoPair = true, the component will automatically create the pairs for each items by cloning the card, just like is now working.

When autoPair = false, is up to the dev to implement manually all the items. The component will verify each id has only 2 elements and report an error otherwise